Module move_ir::parser::ast

Structs

Block

Function

FunctionName

FunctionSignature

IfElse

ImportDefinition

Loop

ModuleDefinition

ModuleName

QualifiedModuleIdent

Script

Spanned

StructDefinition

StructName

StructType
Identifier for a struct definition. Tells us where to look in the storage layer to find the code associated with the interface

Var

While

Enums

BaseExp

BinOp

Builtin
Builtin "function"-like operators that often have a signature not expressable in the type system and/or have access to some runtime/storage context

Cmd

CopyableVal
Bottom of the value hierarchy. These values can be trivially copyable and stored in statedb as a single entry.

Exp

FunctionAnnotation

FunctionBody

FunctionCall

FunctionVisibility

Kind

Statement

Tag

Type

UnaryOp

UseLocal

Functions

format_fields

intersperse

Type Definitions

BaseExp_

Cmd_

CopyableVal_

ExpFields

Exp_

Field

Fields

FunctionCall_

Loc

Var_