Enum move_ir::pinned::ast::BaseVal

pub enum BaseVal {
    CopyableVal(CopyableVal),
    Resource(StructInstance),
    Restricted(StructInstance),
    Array(Vec<BaseVal>),
    Option(Box<Option<BaseVal>>),
}

Lowest of the value hierarchy in the language realm. These values have a literal form and can be stored in structures

Variants
CopyableVal(CopyableVal)Resource(StructInstance)

A linear resource accessible via the methods defined in its struct

Restricted(StructInstance)

A local linear value accessible via the methods defined in its struct Not always copyable as it can hold resources

Array(Vec<BaseVal>)

A container for a vector of value.

Option(Box<Option<BaseVal>>)

An option instance, none or some of a reasource

Methods

impl BaseVal

pub fn infer_type(&self) -> Type

pub fn is_resourceful(&self) -> bool

pub fn address(a: AccountAddress) -> BaseVal

pub fn bool(b: bool) -> BaseVal

pub fn int(i: u64) -> BaseVal

pub fn byte_array(buf: ByteArray) -> BaseVal

pub fn resource_def(o: StructInstance) -> BaseVal

pub fn restricted(o: StructInstance) -> BaseVal

pub fn option_some(v: BaseVal) -> BaseVal

pub fn option_none() -> BaseVal

pub fn instance(n: StructType, s: ValFields) -> BaseVal

pub fn get_access(&self, access: &Access) -> Option<&BaseVal>

pub fn get_mut_access(&mut self, access: &Access) -> Option<&mut BaseVal>

pub fn bind_access_value(
&mut self,
access: Access,
value: BaseVal
) -> Option<()>

pub fn size(&self) -> u64

pub fn from_storage_value(value: &Value) -> BaseVal

pub fn as_storage_value(&self) -> Value